requirements.txt
django
djangorestframework
requests
black